requirements.txt
scrapy
falkordb
google-auth-oauthlib
google-api-python-client
spotipy
dataclasses
